+++ packages/svelte/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.10
 
 - add `paneDistanceClick` prop (max distance between mousedown/up that will trigger a click)
+- returned nodes in `on:nodedragstop` are set to `dragging=false` 
 
 ## 0.1.9